requirements.txt
cachetools==5.3.2
cairocffi==1.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
feedparser==6.0.11
fonttools==4.56.0
google-api-core==2.24.2
google-api-python-client==2.116.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.5
numpy==1.24.4
oauthlib==3.2.2
packaging==24.0
pillow==9.5.0
proto-plus==1.26.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1_modules==0.3.0
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
tinycss2==1.2.1
uritemplate==4.1.1
urllib3==2.0.7
webencodings==0.5.1
